+++ orm/update.go
@@ -11,7 +11,7 @@ func Update(db DB, model interface{}) error {
 	if err != nil {
 		return err
 	}
-	return internal.AssertOneRow(res.Affected())
+	return internal.AssertOneRow(res.RowsAffected())
 }
 
 type updateQuery struct {